feat #17: FeedStore 구현
