+++ PIMO/Sources/View/Home/HomeStore.swift
@@ -12,19 +12,12 @@ import ComposableArchitecture
 
 struct HomeStore: ReducerProtocol {
     struct State: Equatable {
-        var feeds: [Feed] = []
-        var clapCount: Int = 0
+        var feeds: IdentifiedArrayOf<FeedStore.State> = []
     }
     
-    enum Action: BindableAction, Equatable {
-        case binding(BindingAction<State>)
+    enum Action: Equatable {
         case fetchFeeds
-        case moreButtonDidTap
-        case copyButtonDidTap
-        case closeButtonDidTap
-        case clapButtonDidTap
-        case shareButtonDidTap
-        case audioButtonDidTap
+        case feed(id: FeedStore.State.ID, action: FeedStore.Action)
     }
     
     @Dependency(\.homeClient) var homeClient
@@ -32,30 +25,25 @@ struct HomeStore: ReducerProtocol {
     var body: some ReducerProtocol<State, Action> {
         Reduce { state, action in
             switch action {
-            case .binding:
-                return .none
             case .fetchFeeds:
-                state.feeds = homeClient.fetchFeeds()
-            case .moreButtonDidTap:
-                // TODO: 바텀시트
-                let _ = print("more")
-            case .copyButtonDidTap:
-                // TODO: 텍스트 복사
-                let _ = print("copy")
-            case .closeButtonDidTap:
-                // TODO: 텍스트 닫기 (UserDefault)
-                let _ = print("close")
-            case .clapButtonDidTap:
-                state.clapCount += 1
-                let _ = print("clap")
-            case .shareButtonDidTap:
-                // TODO: 딥링크
-                let _ = print("share")
-            case .audioButtonDidTap:
-                // TODO: TTS
-                let _ = print("audio")
+                let feeds = homeClient.fetchFeeds()
+                state.feeds = IdentifiedArrayOf(
+                    uniqueElements: feeds.map { feed in
+                        FeedStore.State(
+                            id: feed.id,
+                            feed: feed,
+                            textImage: feed.textImages[0],
+                            clapButtonDidTap: feed.isClapped
+                        )
+                    }
+                )
+            default:
+                break
             }
             return .none
+        }
+        .forEach(\.feeds, action: /Action.feed(id:action:)) {
+            FeedStore()
         }
     }
 }
